Modules: pass in type parameter to solve type opacity problem
Where the module's functions will need to accept values of arbitrary
type from the outside world, pass in those types as type parameters to
the the trait (module signature) _and_ save the passed-in types
internally under our own names (e.g. T). This is the trick to solving
the problem of type opacity--where the type checker can't prove that the
opaque abstract types defined in traits are the same as the types of the
values that are being passed in to the functions.

But if we _tell_ Scala that they're the same, by passing them in as type
parameters, then no problem--we can have the best of both worlds--type
opacity and easy passing in of values into the modules.