talk: implement modular style
Also rearrange script into separate objects (modules) for the shared
code and the inheritance-based, typeclass-based, and modular
implementations.

Trait companion objects are proving to be a perfect place to stash
functors that create modules that ascribe to the trait signature.

Need to find a way to make the typeclass-based implementation
spiritually similar to the modular one.